extract curve to module level and pull functions out of module pattern
