+++ protocol/src/index.ts
@@ -1,10 +1,6 @@
-// export your SDK here
-
 import {
   AbiCoder,
-  Contract,
   ContractTransactionResponse,
-  EventLog,
   Signer,
   keccak256,
 } from "ethers";
@@ -19,14 +15,14 @@ import {
   generateGroth16Proof,
   toFixedHex,
 } from "./zklib";
-import { ExtPointType, twistedEdwards } from "@noble/curves/abstract/edwards";
-import { Field, mod } from "@noble/curves/abstract/modular";
-import { keccak_256 } from "@noble/hashes/sha3";
+import { ExtPointType } from "@noble/curves/abstract/edwards";
+import { mod } from "@noble/curves/abstract/modular";
 import { randomBytes } from "@noble/hashes/utils";
 import { ensurePoseidon, poseidonHash, poseidonHash2 } from "./poseidon";
 import { bytesToNumberBE } from "@noble/curves/abstract/utils";
 import MerkleTree from "fixed-merkle-tree";
 import { getEncryptionPublicKey } from "@metamask/eth-sig-util";
+import { B, G, getRandomBigInt, R } from "./curve";
 export * from "./config";
 
 export async function outputProve(
@@ -135,7 +131,7 @@ export async function verifySig(
 }
 
 export class MaspContract {
-  constructor(private provider: Signer, private address: string) {}
+  constructor(private provider: Signer, private address: string) { }
 
   async deposit(
     spends: SpendProof[],
@@ -185,61 +181,6 @@ export type SpendProof = {
   valueCommitment: [string, string];
 };
 
-export function getBabyJubJub() {
-  return twistedEdwards({
-    // Params: a, d
-    a: BigInt("168700"),
-    d: BigInt("168696"),
-    // Finite field 𝔽p over which we'll do calculations
-    Fp: Field(
-      BigInt(
-        "21888242871839275222246405745257275088548364400416034343698204186575808495617"
-      )
-    ),
-    // Subgroup order: how many points curve has
-    n: BigInt(
-      "21888242871839275222246405745257275088614511777268538073601725287587578984328"
-    ),
-    // Cofactor
-    h: BigInt(8),
-    // Base point (x, y) aka generator point
-    Gx: BigInt(
-      "995203441582195749578291179787384436505546430278305826713579947235728471134"
-    ),
-    Gy: BigInt(
-      "5472060717959818805561601436314318772137091100104008585924551046643952123905"
-    ),
-    hash: keccak_256,
-    randomBytes,
-  } as const);
-}
-
-export type BabyJub = ReturnType<typeof getBabyJubJub>;
-
-export function getRandomBigInt(bits: number) {
-  const bytes = Math.ceil(bits / 8);
-  const extraBits = bytes * 8 - bits; // Extra bits we get due to byte alignment
-  const arraySize = Math.ceil(bits / 32);
-  const randomValues = new Uint32Array(arraySize);
-  crypto.getRandomValues(randomValues);
-
-  let randomBigInt = BigInt(0);
-  for (let i = 0; i < arraySize - 1; i++) {
-    randomBigInt = (randomBigInt << BigInt(32)) | BigInt(randomValues[i]);
-  }
-
-  // For the last element, only shift the necessary bits
-  randomBigInt =
-    (randomBigInt << BigInt(32 - extraBits)) |
-    (BigInt(randomValues[arraySize - 1]) >> BigInt(extraBits));
-
-  return randomBigInt;
-}
-
-export function getRandomBits(count: number, bits: number) {
-  return new Array(count).fill(0).map(() => getRandomBigInt(bits));
-}
-
 export function toStr(b: bigint): string {
   return "0x" + b.toString(16);
 }
@@ -282,75 +223,63 @@ export async function nullifierHash(
   ]);
 }
 
-export function getInitialPoints(B: BabyJub) {
-  function reddsaSign(a: bigint, A: ExtPointType, msgByteStr: string) {
-    // B - base point
-    // a - secret key
-    // A - Public Key
-    // T - random bytes
-    // M - message bytes
-    // --- sign ------
-    // r = H(T||A||M)
-    // R = r * B
-    // S = r + H(R||A||M) * a
-    // R = H(T||A||M) * B
-    // S = H(T||A||M) + H(R||A||M) * a
-    // --- verify ----
-    // c = H(R||A||M)
-    // -B * S + R + c * A == identity
+function reddsaSign(a: bigint, A: ExtPointType, msgByteStr: string) {
+  // B - base point
+  // a - secret key
+  // A - Public Key
+  // T - random bytes
+  // M - message bytes
+  // --- sign ------
+  // r = H(T||A||M)
+  // R = r * B
+  // S = r + H(R||A||M) * a
+  // R = H(T||A||M) * B
+  // S = H(T||A||M) + H(R||A||M) * a
+  // --- verify ----
+  // c = H(R||A||M)
+  // -B * S + R + c * A == identity
 
-    const abi = new AbiCoder();
-
-    const modN = (a: bigint) => mod(a, B.CURVE.n);
-    const hash = B.CURVE.hash;
-    const BA = B.ExtendedPoint.BASE;
-    const T = randomBytes(32);
-    const r = modN(
-      bytesToNumberBE(
-        hash(
-          abi.encode(
-            ["bytes", "uint256", "uint256", "bytes"],
-            [T, A.x, A.y, msgByteStr]
-          )
-        )
-      )
-    );
-    const R = BA.multiply(r);
-    const cData = abi.encode(
-      ["uint256", "uint256", "uint256", "uint256", "bytes"],
-      [R.x, R.y, A.x, A.y, msgByteStr]
-    );
-    const hashed = keccak256(cData);
-    const c = modN(BigInt(hashed));
-    const s = modN(r + c * a);
-    return { R, s };
-  }
-
-  const [Ro] = getRandomBits(2, 253);
+  const abi = new AbiCoder();
 
   const modN = (a: bigint) => mod(a, B.CURVE.n);
-  const G = B.ExtendedPoint.fromAffine({
-    x: B.CURVE.Gx,
-    y: B.CURVE.Gy,
-  });
-  const R = G.multiply(Ro);
+  const hash = B.CURVE.hash;
+  const BA = B.ExtendedPoint.BASE;
+  const T = randomBytes(32);
+  const r = modN(
+    bytesToNumberBE(
+      hash(
+        abi.encode(
+          ["bytes", "uint256", "uint256", "bytes"],
+          [T, A.x, A.y, msgByteStr]
+        )
+      )
+    )
+  );
+  const R = BA.multiply(r);
+  const cData = abi.encode(
+    ["uint256", "uint256", "uint256", "uint256", "bytes"],
+    [R.x, R.y, A.x, A.y, msgByteStr]
+  );
+  const hashed = keccak256(cData);
+  const c = modN(BigInt(hashed));
+  const s = modN(r + c * a);
+  return { R, s };
+}
 
-  function getV(asset: string) {
-    const V = G.multiply(BigInt(asset));
-    return V;
-  }
+const modN = (a: bigint) => mod(a, B.CURVE.n);
 
-  function valcommit(n: Note) {
-    const r = getRandomBigInt(253);
-    const V = getV(n.asset);
-    const vV =
-      n.amount == 0n ? B.ExtendedPoint.ZERO : V.multiply(modN(n.amount));
-    const rR = R.multiply(modN(r));
-    const Vc = vV.add(rR);
-    return { Vc, r };
-  }
+function getV(asset: string) {
+  const V = G.multiply(BigInt(asset));
+  return V;
+}
 
-  return { G, R, modN, valcommit, getV, reddsaSign };
+function valcommit(n: Note) {
+  const r = getRandomBigInt(253);
+  const V = getV(n.asset);
+  const vV = n.amount == 0n ? B.ExtendedPoint.ZERO : V.multiply(modN(n.amount));
+  const rR = R.multiply(modN(r));
+  const Vc = vV.add(rR);
+  return { Vc, r };
 }
 
 function createNote(amount: bigint, spender: string, asset: string): Note {
@@ -372,7 +301,9 @@ export type Keyset = {
 // Use this to get the public keys from a users private key
 export async function getKeys(privateKey: bigint) {
   await ensurePoseidon();
-  const encryptionKey = getEncryptionPublicKey(privateKey.toString(16).padStart(64, '0'));
+  const encryptionKey = getEncryptionPublicKey(
+    privateKey.toString(16).padStart(64, "0")
+  );
 
   const publicKey = poseidonHash([privateKey]);
 
@@ -383,12 +314,10 @@ export async function getKeys(privateKey: bigint) {
   };
 }
 
-export async function buildMerkleTree(
-  contract: IMasp,
-) {
+export async function buildMerkleTree(contract: IMasp) {
   const filter = contract.filters.NewCommitment();
 
-  const events = await contract.queryFilter(filter)
+  const events = await contract.queryFilter(filter);
   const leaves = events
     .sort((a, b) => {
       return Number(a.args?.index) - Number(b.args?.index);
@@ -425,8 +354,6 @@ async function createProofs(
   sender: Keyset,
   receiver: Keyset
 ) {
-  const babyJub = getBabyJubJub();
-  const { R, modN, valcommit, getV } = getInitialPoints(babyJub);
   const spendProofs: SpendProof[] = [];
   const outputProofs: OutputProof[] = [];
 
@@ -525,11 +452,11 @@ export async function transfer(
 ): Promise<ContractTransactionResponse> {
   logAction(
     "Transferring " +
-      amount +
-      " " +
-      asset +
-      " to " +
-      shrtn(toFixedHex(receiver.publicKey))
+    amount +
+    " " +
+    asset +
+    " to " +
+    shrtn(toFixedHex(receiver.publicKey))
   );
 
   if (signer.provider === null) throw new Error("Signer must have a provider");
@@ -596,6 +523,7 @@ export async function deposit(
     createNote(amount, receiver.publicKey, assetId),
     createNote(0n, receiver.publicKey, assetId),
   ];
+
   const { Bpk, spendProofs, outputProofs } = await createProofs(
     spendList,
     outputList,
